+++ cpp/src/pca/pca_mg.cu
@@ -156,17 +156,18 @@ void fit_impl(raft::handle_t &handle,
     raft::matrix::ratio(handle, explained_var_all.data(),
                         explained_var_ratio_all.data(), prms.n_cols, stream);
 
-    Matrix::truncZeroOrigin(sVector.data(), prms.n_cols, singular_vals,
-                            prms.n_components, 1, stream);
+    raft::matrix::truncZeroOrigin(sVector.data(), prms.n_cols, singular_vals,
+                                  prms.n_components, 1, stream);
 
-    Matrix::truncZeroOrigin(explained_var_all.data(), prms.n_cols,
-                            explained_var, prms.n_components, 1, stream);
-    Matrix::truncZeroOrigin(explained_var_ratio_all.data(), prms.n_cols,
-                            explained_var_ratio, prms.n_components, 1, stream);
+    raft::matrix::truncZeroOrigin(explained_var_all.data(), prms.n_cols,
+                                  explained_var, prms.n_components, 1, stream);
+    raft::matrix::truncZeroOrigin(explained_var_ratio_all.data(), prms.n_cols,
+                                  explained_var_ratio, prms.n_components, 1,
+                                  stream);
 
-    MLCommon::LinAlg::transpose(vMatrix.data(), prms.n_cols, stream);
-    Matrix::truncZeroOrigin(vMatrix.data(), prms.n_cols, components,
-                            prms.n_components, prms.n_cols, stream);
+    raft::linalg::transpose(vMatrix.data(), prms.n_cols, stream);
+    raft::matrix::truncZeroOrigin(vMatrix.data(), prms.n_cols, components,
+                                  prms.n_components, prms.n_cols, stream);
 
     Matrix::opg::deallocate(h, uMatrixParts, input_desc, rank, stream);
 
@@ -207,20 +208,19 @@ void transform_impl(raft::handle_t &handle,
   for (int i = 0; i < input.size(); i++) {
     int si = i % n_streams;
 
-    MLCommon::Stats::meanCenter(input[i]->ptr, input[i]->ptr, mu,
-                                size_t(prms.n_cols), local_blocks[i]->size,
-                                false, true, streams[si]);
+    raft::stats::meanCenter(input[i]->ptr, input[i]->ptr, mu,
+                            size_t(prms.n_cols), local_blocks[i]->size, false,
+                            true, streams[si]);
 
     T alpha = T(1);
     T beta = T(0);
-    LinAlg::gemm(input[i]->ptr, local_blocks[i]->size, size_t(prms.n_cols),
-                 components, trans_input[i]->ptr, local_blocks[i]->size,
-                 int(prms.n_components), CUBLAS_OP_N, CUBLAS_OP_T, alpha, beta,
-                 cublas_h, streams[si]);
+    raft::linalg::gemm(handle, input[i]->ptr, local_blocks[i]->size,
+                       size_t(prms.n_cols), components, trans_input[i]->ptr,
+                       local_blocks[i]->size, int(prms.n_components),
+                       CUBLAS_OP_N, CUBLAS_OP_T, alpha, beta, streams[si]);
 
-    MLCommon::Stats::meanAdd(input[i]->ptr, input[i]->ptr, mu,
-                             size_t(prms.n_cols), local_blocks[i]->size, false,
-                             true, streams[si]);
+    raft::stats::meanAdd(input[i]->ptr, input[i]->ptr, mu, size_t(prms.n_cols),
+                         local_blocks[i]->size, false, true, streams[si]);
   }
 
   if (prms.whiten) {
@@ -311,14 +311,13 @@ void inverse_transform_impl(
     T alpha = T(1);
     T beta = T(0);
 
-    LinAlg::gemm(trans_input[i]->ptr, local_blocks[i]->size,
-                 size_t(prms.n_components), components, input[i]->ptr,
-                 local_blocks[i]->size, prms.n_cols, CUBLAS_OP_N, CUBLAS_OP_N,
-                 alpha, beta, cublas_h, streams[si]);
+    raft::linalg::gemm(handle, trans_input[i]->ptr, local_blocks[i]->size,
+                       size_t(prms.n_components), components, input[i]->ptr,
+                       local_blocks[i]->size, prms.n_cols, CUBLAS_OP_N,
+                       CUBLAS_OP_N, alpha, beta, streams[si]);
 
-    MLCommon::Stats::meanAdd(input[i]->ptr, input[i]->ptr, mu,
-                             size_t(prms.n_cols), local_blocks[i]->size, false,
-                             true, streams[si]);
+    raft::stats::meanAdd(input[i]->ptr, input[i]->ptr, mu, size_t(prms.n_cols),
+                         local_blocks[i]->size, false, true, streams[si]);
   }
 
   if (prms.whiten) {