+++ cpp/src/solver/cd_mg.cu
@@ -110,8 +110,8 @@ void fit_impl(raft::handle_t &handle,
     Matrix::Data<T> squared_data{squared.data(), size_t(input_desc.N)};
     LinAlg::opg::colNorm2NoSeq(squared_data, input_data, input_desc, comm,
                                allocator, streams, n_streams, cublas_handle);
-    LinAlg::addScalar(squared.data(), squared.data(), l2_alpha, input_desc.N,
-                      streams[0]);
+    raft::linalg::addScalar(squared.data(), squared.data(), l2_alpha,
+                            input_desc.N, streams[0]);
   }
 
   std::vector<Matrix::Data<T> *> input_data_temp;
@@ -167,11 +167,12 @@ void fit_impl(raft::handle_t &handle,
         input_data_temp[k]->ptr = input_col_loc;
         input_data_temp[k]->totalSize = partsToRanks[k]->size;
 
-        LinAlg::multiplyScalar(pred_loc, input_col_loc, h_coef[ci],
-                               partsToRanks[k]->size, streams[k % n_streams]);
+        raft::linalg::multiplyScalar(pred_loc, input_col_loc, h_coef[ci],
+                                     partsToRanks[k]->size,
+                                     streams[k % n_streams]);
 
-        LinAlg::add(residual_loc, residual_loc, pred_loc, partsToRanks[k]->size,
-                    streams[k % n_streams]);
+        raft::linalg::add(residual_loc, residual_loc, pred_loc,
+                          partsToRanks[k]->size, streams[k % n_streams]);
 
         pred_loc = pred_loc + partsToRanks[k]->size;
         residual_loc = residual_loc + partsToRanks[k]->size;
@@ -209,11 +210,12 @@ void fit_impl(raft::handle_t &handle,
       for (int k = 0; k < input_data.size(); k++) {
         input_col_loc = input_data[k]->ptr + (ci * partsToRanks[k]->size);
 
-        LinAlg::multiplyScalar(pred_loc, input_col_loc, h_coef[ci],
-                               partsToRanks[k]->size, streams[k % n_streams]);
+        raft::linalg::multiplyScalar(pred_loc, input_col_loc, h_coef[ci],
+                                     partsToRanks[k]->size,
+                                     streams[k % n_streams]);
 
-        LinAlg::subtract(residual_loc, residual_loc, pred_loc,
-                         partsToRanks[k]->size, streams[k % n_streams]);
+        raft::linalg::subtract(residual_loc, residual_loc, pred_loc,
+                               partsToRanks[k]->size, streams[k % n_streams]);
 
         pred_loc = pred_loc + partsToRanks[k]->size;
         residual_loc = residual_loc + partsToRanks[k]->size;
@@ -313,13 +315,13 @@ void predict_impl(raft::handle_t &handle,
 
   for (int i = 0; i < input_data.size(); i++) {
     int si = i % n_streams;
-    LinAlg::gemm(input_data[i]->ptr, local_blocks[i]->size, input_desc.N, coef,
-                 preds[i]->ptr, local_blocks[i]->size, size_t(1), CUBLAS_OP_N,
-                 CUBLAS_OP_N, alpha, beta, handle.get_cublas_handle(),
-                 streams[si]);
+    raft::linalg::gemm(handle, input_data[i]->ptr, local_blocks[i]->size,
+                       input_desc.N, coef, preds[i]->ptr, local_blocks[i]->size,
+                       size_t(1), CUBLAS_OP_N, CUBLAS_OP_N, alpha, beta,
+                       streams[si]);
 
-    LinAlg::addScalar(preds[i]->ptr, preds[i]->ptr, intercept,
-                      local_blocks[i]->size, streams[si]);
+    raft::linalg::addScalar(preds[i]->ptr, preds[i]->ptr, intercept,
+                            local_blocks[i]->size, streams[si]);
   }
 }